feature: cross target testing
- adding cross compile and test targets to crucible
- read from repo's config file for enabling fuzz and cross
