+++ .crucible.toml
@@ -1,0 +1,2 @@
+[cross]
+enable = false
