+++ src/common/mod.rs
@@ -2,3 +2,4 @@ pub mod config;
 pub mod logging;
 pub mod metrics;
 pub mod options;
+pub mod repoconfig;
